imp: added re-init for migration
